# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Tesis Guvenlik & Operasyon API
  version: "0.1.0"
  description: |
    Multi-tenant tesis guvenlik & operasyon SaaS — v0 sozlesmesi.
    Backend (FastAPI) ile Mobil (Flutter) ve Panel (Next.js) gelistiricilerinin
    paralel calismasi icin tek dogruluk kaynagidir.

    Konvansiyonlar (detay: /contracts/README.md):
      * Tum zamanlar ISO8601 UTC (`...Z`). tenant.timezone ile yorumlanir.
      * Hata zarfi: `{ "error": { "code": "...", "message": "..." } }`.
      * Sayfalama: `limit` (varsayilan 50, max 200) + `offset`. Liste yanitlari
        `PageMeta` icerir.
      * Auth: JWT Bearer (bkz. /contracts/auth.md). Tenant her zaman token'dan
        cikarilir — istemci tenant_id gondermez/secemez.
      * `POST /scans` icin `Idempotency-Key` header zorunlu.
servers:
  # NOT: base path /v0 KALDIRILDI — gercek backend endpoint'leri /v0 ONEKI OLMADAN
  # sunuyor (orn. /auth/login, /scans). Base = host:port.
  - url: https://api.example.com
    description: production
  - url: http://localhost:8000
    description: yerel gelistirme
...
security:
  - bearerAuth: []
paths:
...
  /tasks:
    get:
      tags: [tasks]
      summary: Gorev listesi
      description: |
        `atanan_user_id` filtresi (mobil §11 #1): `me` = token kullanicisina
        atananlar ("Gorevlerim" tek istekte); duz UUID = o kullaniciya atananlar
        (panel; tum okuyucu roller kullanabilir — atama bilgisi zaten listede
        gorunur, yetki siniri degildir). Gecersiz deger 422.
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: tip
          in: query
          schema: { $ref: "#/components/schemas/TaskTip" }
        - name: aktif
          in: query
          schema: { type: boolean }
        - name: atanan_user_id
          in: query
          description: "'me' veya user UUID — atanan kullanici filtresi"
          schema: { type: string }
      responses:
        "200":
          description: Liste
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    required: [items]
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/Task" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [tasks]
      summary: Gorev olustur
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaskCreate" }
      responses:
        "201":
          description: Olusturuldu
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /tasks/{id}:
    parameters:
      - $ref: "#/components/parameters/PathId"
    get:
      tags: [tasks]
      summary: Gorev detay
      responses:
        "200":
          description: Detay
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [tasks]
      summary: Gorev guncelle (kismi)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaskUpdate" }
      responses:
        "200":
          description: Guncellendi
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationError" }
    delete:
      tags: [tasks]
      summary: Gorev sil
      responses:
        "204": { description: Silindi }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /tasks/{id}/completions:
    parameters:
      - $ref: "#/components/parameters/PathId"
    get:
      tags: [tasks]
      summary: Gorev tamamlama gecmisi
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Liste
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    required: [items]
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/TaskCompletion" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [tasks]
      summary: Gorevi tamamla (kanit gonder)
      description: |
        Idempotency-Key header ZORUNLU (offline cift gonderim korumasi, scan deseni).
        Foto icin: once POST /uploads/presign ile foto_key + PUT URL alinir, dosya
        dogrudan MinIO'ya yuklenir, sonra burada foto_key gonderilir.

        FOTO ZORUNLULUGU (mobil §11 #2): gorevde `foto_zorunlu=true` ise
        `foto_key` OLMADAN tamamlama **422** doner (anlamli mesajla).

        NFC eslesmesi NORMALIZE (mobil §11 #3): gorevin checkpoint'i varsa
        gonderilen `nfc_tag_uid` strip + upper sonrasi karsilastirilir
        (buyuk/kucuk harf, bas/son bosluk farki 422 dusurmez).
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaskCompletionCreate" }
      responses:
        "201":
          description: Tamamlama kaydedildi
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskCompletion" }
        "200":
          description: Idempotent tekrar — mevcut kayit
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskCompletion" }
        "400":
          description: Idempotency-Key eksik veya gecersiz govde
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Ayni Idempotency-Key farkli govde ile gonderildi
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
  /tenant/settings:
    get:
      tags: [tenant]
      summary: Tenant ayarlari (acil durum yonetim numarasi dahil)
      description: "Mobil, acil durumda aranacak `acil_durum_telefon`'u buradan okur (tel: link)."
      responses:
        "200":
          description: Ayarlar
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TenantSettings" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    patch:
      tags: [tenant]
      summary: Tenant ayarlarini guncelle (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TenantSettingsUpdate" }
      responses:
        "200":
          description: Guncellendi
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TenantSettings" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationError" }

  # ================================================================= #
  # AIDAT — unit (daire) + sakin + tahakkuk + odeme
  # ================================================================= #
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access token. `Authorization: Bearer <token>`. Bkz. /contracts/auth.md"
  parameters:
    Limit:
      name: limit
      in: query
      description: Sayfa boyutu (varsayilan 50, max 200)
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Offset:
      name: offset
      in: query
      description: Atlanan kayit sayisi
      schema: { type: integer, minimum: 0, default: 0 }
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Istemci tarafindan uretilen benzersiz anahtar (UUID onerilir)
      schema: { type: string, minLength: 8, maxLength: 200 }
  responses:
    BadRequest:
      description: Gecersiz istek
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Kimlik dogrulama gerekli / gecersiz token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Rol bu islem icin yetkili degil
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Kayit bulunamadi (tenant kapsaminda)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ValidationError:
      description: Govde dogrulama hatasi
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Makine-okunabilir hata kodu
              example: validation_error
            message:
              type: string
              description: Insan-okunabilir aciklama
              example: "baslangic_saat bitis_saat'ten once olmali"
            details:
              type: array
              description: Alan bazli dogrulama hatalari (opsiyonel)
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
    Page:
      type: object
      required: [meta]
      properties:
        meta: { $ref: "#/components/schemas/PageMeta" }
    PageMeta:
      type: object
      required: [limit, offset, total]
      properties:
        limit: { type: integer, example: 50 }
        offset: { type: integer, example: 0 }
        total:
          type: integer
          description: Filtreye uyan toplam kayit sayisi
          example: 137
...
    TaskTip:
      type: string
      enum: [temizlik, kontrol, ilaclama, bakim, peyzaj, diger]
    Task:
      type: object
      required: [id, tip, ad, foto_zorunlu, aktif, created_at]
      properties:
        id: { type: string, format: uuid }
        tip: { $ref: "#/components/schemas/TaskTip" }
        ad: { type: string, example: "Cop toplama" }
        aciklama: { type: string, nullable: true }
        atanan_user_id: { type: string, format: uuid, nullable: true }
        checkpoint_id: { type: string, format: uuid, nullable: true }
        periyot_dakika:
          type: integer
          minimum: 1
          nullable: true
          description: "Tekrar araligi (dakika); tek seferlikse null"
        sonraki_planlanan:
          type: string
          format: date-time
          nullable: true
          description: "Bir sonraki planlanan an (UTC); peyzaj takvimi. Tamamlaninca ilerler."
        foto_zorunlu:
          type: boolean
          description: "true ise completion foto_key olmadan kabul edilmez (422) — mobil §11 #2"
        aktif: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    TaskCreate:
      type: object
      required: [tip, ad]
      properties:
        tip: { $ref: "#/components/schemas/TaskTip" }
        ad: { type: string }
        aciklama: { type: string, nullable: true }
        atanan_user_id: { type: string, format: uuid, nullable: true }
        checkpoint_id: { type: string, format: uuid, nullable: true }
        periyot_dakika: { type: integer, minimum: 1, nullable: true }
        sonraki_planlanan: { type: string, format: date-time, nullable: true }
        foto_zorunlu: { type: boolean, default: false }
        aktif: { type: boolean, default: true }
    TaskUpdate:
      type: object
      minProperties: 1
      properties:
        tip: { $ref: "#/components/schemas/TaskTip" }
        ad: { type: string }
        aciklama: { type: string, nullable: true }
        atanan_user_id: { type: string, format: uuid, nullable: true }
        checkpoint_id: { type: string, format: uuid, nullable: true }
        periyot_dakika: { type: integer, minimum: 1, nullable: true }
        sonraki_planlanan: { type: string, format: date-time, nullable: true }
        foto_zorunlu: { type: boolean }
        aktif: { type: boolean }
    TaskCompletion:
      type: object
      required: [id, task_id, tamamlayan_user_id, tamamlanma_zamani, idempotency_key, created_at]
      properties:
        id: { type: string, format: uuid }
        task_id: { type: string, format: uuid }
        tamamlayan_user_id:
          type: string
          format: uuid
          description: "Token'dan turetilir"
        tamamlanma_zamani: { type: string, format: date-time }
        nfc_tag_uid: { type: string, nullable: true }
        gps_lat: { type: number, format: double, nullable: true }
        gps_lng: { type: number, format: double, nullable: true }
        foto_key: { type: string, nullable: true }
        foto_url: { type: string, nullable: true }
        notlar: { type: string, nullable: true }
        idempotency_key: { type: string }
        created_at: { type: string, format: date-time }
    # Capraz-gorev tamamlama gecmisi ogesi (GET /task-completions). foto_url/gps
    # YOK; kanit varligi foto_var/nfc_dogrulandi bool olarak yeter.
...
    TaskCompletionCreate:
      type: object
      required: [tamamlanma_zamani]
      properties:
        tamamlanma_zamani: { type: string, format: date-time }
        nfc_tag_uid:
          type: string
          nullable: true
          description: "Verilirse task.checkpoint_id'nin nfc'si ile eslesmeli"
        gps_lat: { type: number, format: double, nullable: true }
        gps_lng: { type: number, format: double, nullable: true }
        foto_key:
          type: string
          nullable: true
          description: "Once /uploads/presign ile alinan obje anahtari"
        notlar: { type: string, nullable: true }

    # ---------------------------- uploads -------------------------- #
...
    TenantSettings:
      type: object
      required: [tenant_id, ad, slug, timezone]
      properties:
        tenant_id: { type: string, format: uuid }
        ad: { type: string }
        slug: { type: string }
        timezone: { type: string }
        acil_durum_telefon:
          type: string
          nullable: true
          description: "Acil durumda mobilin arayacagi yonetim numarasi (tel:)"
    TenantSettingsUpdate:
      type: object
      minProperties: 1
      properties:
        acil_durum_telefon: { type: string, nullable: true }
        timezone: { type: string }
        ad: { type: string }

    # ----------------------------- aidat --------------------------- #
...